Change Namespace into NamespaceTree
`Namespace` becomes `NamespaceTree` and to better support the async and
nested `RemoteData` (the intent being to do a request to the server when
expanding a node in the tree).

Additionally clean up a few things: Move common view functions into a
`UI` module, rename `UnisonHash` to `Hash`, and remove `Path`
functionality from `NamespaceTree`.