Add FQN and expandedNamespaces state
Add the `FullyQualifiedName` module with the `FQN` type and parse JSON
names into this (the backend does not yet serve an FQN and the parsing here
is a bit presumptious).

Additionally add a `Set` of `FQN`s to the App `Model` for keeping track
of which namespaces are expanded (This could, later on, be serialized in
the URL and drive which namespaces are fetched from a "cold start").