Add HashSet to track expandedNamespaces by Hash
Add `HashSet` as a small wrapper around `Set` and `Hash` to allow for
easy manipulation of the wrapped `Hash` type when in sets (`Hash` isn't
"comparable" in the eyes of Elm natively).

Use this `HashSet` to drive the `expandedNamespaces` state and update
the `update` function to handle toggling of namespace hashes in that
state.