Add OpenDefinitions type and initial fetch
Add a new type: `OpenDefinitions` that tracks the state of the
`Definitions` open in the workspace area. It's an opaque wrapper around
an `OrderedMap` from `Hash` `WebData Definition`.

Add `OpenDefinitions` to `App` and use it to fetch `Definitions` by
`Hash`, render a loading indicator for each Definition being fetched
(correctly positioned due to the wrapped `OrderedMap`) and render a
placeholder for when the data is returned from the server.

Note that the `Definition` decoders are placeholder as the API is
changing shortly.

Additionally clean up the UI and Icons a little and make `Hash` opaque.