Add Env file and update syntax tags
The backend changed to be on a random port and to generate a token at
start-up. Add a new build step to generate an `Env.elm` file based on a
token passed into the `npm start` script, which also now requires an API
url. The full start command is now:

```sh
API_URL=<API URL FROM UCM> API_TOKEN=<API TOKEN FROM UCM> npm start
```

This `Env.elm` is generated with a small node script - The intent is
that it will eventually support environment settings from something like
[`dotenv`](https://github.com/motdotla/dotenv).

A generate file is choose instead of `flags` to the `init` function to
make it easier to deal with these static settings without having to pass
them around as state in Elm.

--

Additionally update the Syntax module to support a few backend renames
around references and hash qualifiers.