Fuzzy Finder: Render definition matches
Render (mocked for now) definition matches based on the SearchResults
data structure, supporting navigation with arrow keys and mouse over and
selection with the enter key and mouse click, and indicating both focus
and keyboard shortcuts (the sequence shortcuts are not yet implemented).

The width of the definition name column name will vary depending on the
max character width of the names in the search result, resulting in the
signature being as close as possible to the left side while still being
aligned.

Introduce a new KeyboardShortcuts module for displaying shortcuts (and
eventually handling things like sequence and combos), setting up events
and decoding keyCodes into keys (using third-party libraries).

Also update the SearchResults data structure to be a little easier to
work with.