Add WorkspaceItems to replace OpenDefinitions
Add in the types and tests related to `WorkspaceItems` such that it can
eventually replace `OpenDefinitions`.

`WorkspaceItems` differs from `OpenDefinitions` in that its `Reference`
based instead of `Hash` based and supports `Term` and `Type` directly
instead of through `Definition` (which will be retired in its current
form). It also differs in that it tracks `Loading`, `Failure` and
`Success` directly instead of through `RemoteData`.

`Term` and `Type` are now their own thing that share an `Info` between
them. They are somewhat extensible allowing for both summary levels of
data (for things like the `Finder`), listing data for
`NamespaceListings` and detailed versions for the `Workspace`.

Note: that none of this is yet in use.