Fix namespaceOf bug when suffix appears many times
When a suffixName is part of an FQN multiple times, say "Map" being part
of "base.Map.Map" there was a bug where all "Map" occurrences were being
removed resulting in "base.". Instead of just replacing the occurrences,
simply drop the number of characters in the suffix from the right of the
string.