Build tooling: Setup Webpack
Static files are moved to the src directory such that they are included
in the bundle when webpack builds.

A minor snag occurred with svg icons in that they were rendered lower in
the page than the non webpack version. Adding webpack shouldn't have had
an impact on this, but adding a font-size to the svg wrapping element
fixed it.