Add top level support for Constructors
Add support for Constructors (either Ability- or DataConstructors) on
the same level as Terms and Types. They are related to both (technically
are Terms), but have some unique characteristics that is more
ergnomically handled in the UI if they are handled as they own unique
thing.

This also adds routing support for constructor and support for icons
(though this commit just uses type and ability icons for them).