Bugfix: previously fetched listings were blank
When expanding a namespace and then expanding a subnamespace within and
then collapsing the original namespace and expanding it again, the
subnamespace would appear in an "expanded" state with the down caret,
but there would be no content shown within and the only way to see the
content would be to collapse and re-expand.

Add a `contentFetched` function to the NamespaceListing to solve this
problems such that when expanding, re-fetching (and overwriting)
previously expanded namespaces doesn't happen —  meaning only namespaces
that haven't been fetched before are fetched.

This fixes https://github.com/unisonweb/codebase-ui/issues/47