Support running the UI from ucm
Support static assets being served by ucm on the root (like
`/static/favicon.svg`), and the UI index file and API served on the
dynamic `/:token/ui` and `/:token/api` respectively.

Update the webpack config to work with the `<base>` html tag to support
relative routing with a base path while keeping assets served from an
absolute path. Pass the base path and api base path to the Elm app to
ensure routes and api requests are prefixed appropriately.

Update the Api to require the apiBasePath as late in the creation as
possible to easy the ergonimics of api requests.

Rename System to Env (as the autogenerated Env file is no longer
needed).