Add a loader to Finder and improve error messages
When the Finder takes more than ~250ms it starts to get visible and the
user is left in an awkward state when there's no feedback. Remedy this
by adding a animation on the magnifying glass icon and a transparent
overlay over the previous results. This also meant using a specialized
data structure for the finder search (named FinderSearch) over
RemoteData such that a search on top of a previously successful search
result could show the old result (albeit dimmed) while loading the next
result.

Additionally improve rendering of various error messages across the app.