Add search icon animation; Redo icons.
To support search icon animation (and animation for icons in general),
pull in all icons (some with slightly rounder outlines) into Elm. This
allows in-depth styling, animation, and is included in the js bundle
such that the icons can be cached.

Animations weren't possible with the previous file based svg sprite
sheets, but could have been done by embedding the svg into the html
document. This opts for using Elm to produce the icons directly for
maximum control and cacheability. The downside being an increased
document size (each instance of an icon in the page is added to the
document in full).