Debounce Finder query for Unison Share
Add a debounce in the Finder when searching for definitions from Unison
Share such that the server isn't hit hard with requests on each
keystroke that are since thrown away by the UI since they are not
current with the search input. Set the debounce to 300ms.

Don't debounce for Ucm.