Add folding support to Docs
Track folding state for the `Doc` elements `Folded` and `FoldedSource`
through a `DocFoldedToggles` type that wraps a `Set` of a new `Id` type
(a simple `Id` type with a phantom type for type safe ids). Use this new
new folding state on `WorkspaceItem`, which has gotten most of its
fields merged in to a record to make it easier to add more in the future
and to have them be named.

Add a `Id Doc` field to `Folded` and `FoldedSource` (intended to be
generated on parse time) for identifying which node in the doc tree has
its fold toggled.

Toggling works by being either present or not. When present, it is the
opposite of the original fold state that was provided by the
server—meaning the state that was in the original doc the author wrote.

Add a few stylistic improvements to folding both in docs and in the main
sidebar such that the expand doesn't animate, but the caret does and the
caret in the doc fold has been given more space such that it is an
easier click target.