Add a button in the sidebar to change perspective
Add a small button on the right side of the sidebar on namespace items
that allows users to change the perspective to that namespace.

To support this add a new notion of buttons: "contained": regular
looking buttons, and "uncontained": buttons that have a transparent
background until hovered, a pattern seen most often in Google's Material
UI Design System.