Use root relative FQNs in CodebaseTree perspective
Fix a bug that would use a relative to the current NamespacePerspective
FQN when setting the perspective to a sub namespace of the current
perspective, causing the codebase tree to appear blank (because it
requested a namespace listing with a partial FQN).

Fix this by appending the current namespace perspective FQN with the new
one being set.