Enable folding definition source
Add a new FoldToggle UI module to consolidate various folding mechanisms
and use it in docs, source, and the main definition fold.

With this definition source can now be folded and is part of the Zoom
cycle. Most definitions start with their source folded open, except for
docs that start with it closed.

Also clean up how margins are handled for first and last elements in
docs.