Fix: When changing perspective, the tree is blank
Fix a bug that happened when changing the perspective and resulting the
in codebase tree in the sidebar to be blank right after. This was due
to both the change perspective click and the expand namespace events
were fired and the bug would happened conditionally on which request
would finish first.

Fix this by adding stopPropagation to the change perspective click via
extensions to the Button module to support configurations of this kind.