Crop long definition docs
When definition docs are very long (and the term isn't a doc itself),
crop the doc and provide an affordance to show the full doc.

To determine if the doc is very long, render it in an `overflow: hidden`
container and query the dom to see if that container has content that is
cropped.