Fix Safari Finder render issue
In Safari (and perhaps other browsers?) `transparent` when used in
gradients are rendered with a strange dark fade instead of actually
being transparent. Using rgba with a 0 alpha fixes this. Specifically
this was noticed in the Finder (there's a truncation fade effect on the
right hand side), but could be else where in the app as well.

The fix was to add a new css variable: `--color-transparent` that
evaluates to `rgba(255, 255, 255, 0)` and use that in place of
`transparent` everywhere.