Improve the rendering of FQNs
Render FQNs with a more subtle and spacious treatmen to the separator
between segments. Also prefer the shortest FQN for picking the namespace
of an suffixed FQN. This mitigates some of the weird cases where forks
were of libraries were picked over the actual libraries, but wouldn't
work if the length of the segments happen to be equal (longer term
solution is to pick the FQN that was used to query the definition).

Also adjusted the Find Definition button slightly in size to match
latest designs