Organize the routing model to favor "latest"
Ensure that we use Relative everywhere and disable Absolute URLs until
we are ready to support them. Add comments to communicate why this is
done (Share doesn't support history) and change PreApp to fetch the
perspective information from Api.codebaseHash instead of Api.list (so we
can limit Api to only use Perspective instead of PerspectiveParams).