Pin open definitions to hash on perspective change
When there's any number of definitions open and a new perspective is
selected, the definitions that were open with references based on names,
will no longer have a resolvable name if they were to be re-queried (on
a page refresh for instance). Fix this by changing all open definitions
to be referenced by hash when the perspective is changed.