Ensure Definitions are opened via Route changes
We want the URL to be main way definitions are opened and fetched.
Previously this was an afterthought such that we fetched a definition
and then changed the URL, resulting in broken back-button behavior.

This changes how we load definitions such that a URL is changed which,
in a route handler, causes the Workspace to open and fetch that
definition, with 1 exception: when a definition is opened from another
definition, it is important that we open the new definition relative to the
originating definition (above it or below it). In this case we open the
definition, then change the URL, which subsequently tries to open the
definition again; seeing that the definition has been opened already the
fetch attempt is cancelled.