Catalog: Fetch catalog doc and projects and render
Fetch the full catalog for the Catalog page by fetching the special
catalog doc and parsing it into a CatalogMask which is used to
categorize projects which are now also fetched.

With the data in hand the page renders a series of `UI.Card` components
for each category with their projects.

Note that the categories are not currently sorted, but we'll want them
to be in the future.