Temporarily hide the _catalog doc
Unison Share will have a new front page that features a catalog which is
derived from a Unison Doc. This page will mean that users will never see
a tree view of the top of the Unison Share like they do now.

While the Catalog page is being worked on, but not switched over to, we
want to hide the _catalog doc from the top level codebase tree to avoid
any confusion. This should be removed later when the Catalog page is in
place.