Move route specific keyboardshortcuts to pages
Ensure that when we are on the Catalog say, we don't trigger Workspace
keyboardshortcuts.

Notably this also means that things like the Finder are no longer an App
concern, by a Workspace shortcut. There's a few other things that should
be moved down a layer like toggling the sidebar, but this should only
happen for Share and not for Local (this is not tackled in this effort).